+++ new-client/src/plugins/DocumentHandler/utils/htmlToMaterialUiParser.js
@@ -1,3 +1,5 @@
+import { v4 as uuidv4 } from "uuid";
+
 /**
  * Default export of function that takes a html-string as input and returns array with MaterialUIComponents
  * @param {string} html String with html that needs to be converted to materialUIComponents
@@ -24,6 +26,7 @@ const parseChild = (child, tagSpecificCallbacks) => {
     return object.tagType.toLowerCase() === child.tagName.toLowerCase();
   });
   child.callback = callbackObject.callback;
+  child.componentId = uuidv4();
   if (child.children.length > 0) {
     for (const subChild of child.children) {
       parseChild(subChild, tagSpecificCallbacks);
